eager import trio by shoving worker behavior into another top level package
